--- allennlp/data/dataloaders/__init__.py
@@ -1,8 +1,0 @@
-from allennlp.data.dataloaders.dataloader import DataLoader, TensorDict, allennlp_collate
-from allennlp.data.dataloaders.multi_process_dataloader import MultiProcessDataLoader
-from allennlp.data.dataloaders.pytorch_dataloader import (
-    PyTorchDataLoader,
-    allennlp_worker_init_fn,
-    AllennlpDataset,
-    AllennlpLazyDataset,
-)
