+++ tests/commands/evaluate_test.py
@@ -8,7 +8,7 @@ import pytest
 
 from allennlp.commands.evaluate import evaluate_from_args, Evaluate, evaluate
 from allennlp.common.testing import AllenNlpTestCase
-from allennlp.data.dataloaders import TensorDict
+from allennlp.data.data_loaders import TensorDict
 from allennlp.models import Model
